+++ Domain/ValueData.cs
@@ -8,6 +8,6 @@ public class ValueData
     [Required]
     public string Name { get; set; } = string.Empty;
     public DateTime Date { get; set; }
-    public decimal ExecutionTime { get; set; }
-    public decimal Value { get; set; }
+    public double ExecutionTime { get; set; }
+    public double Value { get; set; }
 }
