+++ Models/ResultFilterDTO.cs
@@ -7,11 +7,11 @@ public class ResultFilterDTO
     public DateTime? MinFirstRecordTime { get; set; }
     public DateTime? MaxFirstRecordTime { get; set; }
 
-    public decimal? MinAverageValue { get; set; }
-    public decimal? MaxAverageValue { get; set; }
+    public double? MinAverageValue { get; set; }
+    public double? MaxAverageValue { get; set; }
 
-    public decimal? MinAverageExecutionTime { get; set; }
-    public decimal? MaxAverageExecutionTime { get; set; }
+    public double? MinAverageExecutionTime { get; set; }
+    public double? MaxAverageExecutionTime { get; set; }
 
     public int PageNumber { get; set; } = 1;
     public int PageSize { get; set; } = 10;
